refactor(chat): own durable index state together
